moved PlayerCamera to graphics package; removed currentState from GameManager; built in time out in millis in PlayerCamera
skrrr
